refactor: shifting source files with dir to be `index.ts`
