+++ src/index.ts
@@ -1,2 +1,2 @@
-export { createTestCppSolutionTasks } from "./test/cpp.js";
+export { createTestCppSolutionTasks } from "./test/cpp/index.js";
 export { readYamlSchema, Schema } from "./test/schema.js";
