+++ src/test/cpp/generate/index.ts
@@ -1,9 +1,9 @@
 import { mkdirSync, writeFileSync } from "node:fs";
 import path from "node:path";
-import { Schema } from "../schema.js";
-import { generateCppMainCode } from "./generate/main.js";
-import { generateCppUtilityCode } from "./generate/utility.js";
-import { generateCppTestCaseCode } from "./generate/test_case.js";
+import { Schema } from "../../schema.js";
+import { generateCppMainCode } from "./main.js";
+import { generateCppUtilityCode } from "./utility.js";
+import { generateCppTestCaseCode } from "./test_case.js";
 
 /**
  * Generates a C++ test file from a test schema.
